convert normal tests
